Move logout features out of masthead menu
